Scanner & Compiler dummies

+++ Sources/DenrimScript/OpCode.swift
@@ -5,7 +5,7 @@
 //  Created by Markus Moenig on 5/10/2564 BE.
 //
 
-typealias OpCodeType = UInt16
+typealias OpCodeType = UInt8
 
 enum OpCode : OpCodeType {
